Necessary types moved to Application.Infrastructure
