+++ src/CaptainHook.Api/Controllers/SubscribersController.cs
@@ -1,10 +1,9 @@
 ﻿using System;
 using System.Threading.Tasks;
 using CaptainHook.Api.Constants;
-using CaptainHook.Application.Gateways;
+using CaptainHook.Application.Infrastructure.DirectorService;
 using CaptainHook.Common;
 using CaptainHook.Common.Configuration;
-using CaptainHook.Common.Remoting;
 using Eshopworld.Core;
 using MediatR;
 using Microsoft.AspNetCore.Authorization;
